LMS/requirements.txt
celery>=5.2.7
django-celery-results>=2.4.0
amqp>=5.1.1  # AMQP libraries for RabbitMQ